refactor: rename base-url
